Revert "Merge branch 'master' into IT47_Jovan_Milosevic"
This reverts commit ab5d0d0febf671c9dab1f18deff927654f3acbfb, reversing
changes made to 80708300b0368d8b1de3317e4fd99a75a3b63c44.

+++ src/Business/Grand.Business.Checkout/Commands/Handlers/Orders/OrderStatusCommandHandler.cs
@@ -127,7 +127,7 @@ namespace Grand.Business.Checkout.Commands.Handlers.Orders
 
             //gift vouchers activation
             if (_orderSettings.GiftVouchers_Activated_OrderStatusId > 0 &&
-               _orderSettings.GiftVouchers_Activated_OrderStatusId == request.Order.OrderStatusId)
+               _orderSettings.GiftVouchers_Activated_OrderStatusId == (int)request.Order.OrderStatusId)
             {
                 await _mediator.Send(new ActivatedValueForPurchasedGiftVouchersCommand() { Order = request.Order, Activate = true });
             }
